Exclude AssimpNet/win DLLs from WebGL builds
In the same project, we use URDF-Importer in Unitye Editor, but we also have scenes
that are built for WebGL.

As noted in https://github.com/Unity-Technologies/URDF-Importer/issues/201
WebGL builds of a project that has URDF-Importer package always fail, even if
URDF-Importer is not used during runtime.

This commit marks the conflicting plugins to exclude WebGL.
NOTE: URDF-Importer scripts will not be usable in WebGL builds: Runtime/Unity.Robotics.URDFImporter.asmdef
still excludes WebGL. This change allows scenes that don't use URDF-Importer to build.